--- tests/Common/LegacyTokensTest.php
@@ -1,396 +1,0 @@
-<?php
-/**
- *
- * User: Martin Halamíček
- * Date: 16.5.12
- * Time: 11:46
- *
- */
-
-namespace Keboola\Test\Common;
-
-use Keboola\StorageApi\ClientException;
-use Keboola\Test\StorageApiTestCase;
-use Keboola\Csv\CsvFile;
-use Keboola\StorageApi\Options\Components\ListComponentsOptions;
-use Keboola\StorageApi\Client;
-
-class LegacyTokensTest extends StorageApiTestCase
-{
-
-    protected $_inBucketId;
-    protected $_outBucketId;
-
-    public function setUp()
-    {
-        parent::setUp();
-        $this->_initEmptyTestBuckets();
-
-        $triggers = $this->_client->listTriggers();
-        foreach ($triggers as $trigger) {
-            $this->_client->deleteTrigger((int) $trigger['id']);
-        }
-
-        $this->_outBucketId = $this->getTestBucketId(self::STAGE_OUT);
-        $this->_inBucketId = $this->getTestBucketId(self::STAGE_IN);
-
-        $this->_initTokens();
-    }
-
-    protected function _initTokens()
-    {
-        foreach ($this->_client->listTokens() as $token) {
-            if ($token['isMasterToken']) {
-                continue;
-            }
-            $this->_client->dropToken($token['id']);
-        }
-    }
-
-    protected function _clearComponents()
-    {
-        $components = new \Keboola\StorageApi\Components($this->_client);
-        foreach ($components->listComponents() as $component) {
-            foreach ($component['configurations'] as $configuration) {
-                $components->deleteConfiguration($component['id'], $configuration['id']);
-            }
-        }
-
-        // erase all deleted configurations
-        foreach ($components->listComponents((new ListComponentsOptions())->setIsDeleted(true)) as $component) {
-            foreach ($component['configurations'] as $configuration) {
-                $components->deleteConfiguration($component['id'], $configuration['id']);
-            }
-        }
-    }
-
-    public function testTokenComponentAccess()
-    {
-
-        $this->_clearComponents();
-
-        $description = "Component Access Test Token";
-        $componentAccess = array("wr-db");
-        $bucketPermissions = array($this->_inBucketId => "write");
-
-        $componentAccessTokenId = $this->_client->legacyCreateToken($bucketPermissions, $description, null, false, $componentAccess);
-        $componentAccessToken = $this->_client->getToken($componentAccessTokenId);
-
-        $componentFailTokenId = $this->_client->legacyCreateToken($bucketPermissions, $description);
-        $componentFailToken = $this->_client->getToken($componentFailTokenId);
-
-        $accessClient = new \Keboola\StorageApi\Client(array(
-            'token' => $componentAccessToken['token'],
-            'url' => STORAGE_API_URL
-        ));
-
-        $failClient = new \Keboola\StorageApi\Client(array(
-            'token' => $componentFailToken['token'],
-            'url' => STORAGE_API_URL
-        ));
-
-        // we'll test 3 scenarios, using an admin token, a token with some component access, and a token with no access
-        $adminComponentClient = new \Keboola\StorageApi\Components($this->_client);
-        $accessComponentClient = new \Keboola\StorageApi\Components($accessClient);
-        $failComponentClient = new \Keboola\StorageApi\Components($failClient);
-
-        // we should have no components at the start
-        $allComponents = $adminComponentClient->listComponents();
-        $this->assertCount(0, $allComponents);
-
-        // we'll create some initial test configuration
-        $adminComponentClient->addConfiguration((new \Keboola\StorageApi\Options\Components\Configuration())
-            ->setComponentId('wr-db')
-            ->setConfigurationId('main-1')
-            ->setName('Main1'));
-        $adminComponentClient->addConfiguration((new \Keboola\StorageApi\Options\Components\Configuration())
-            ->setComponentId('wr-db')
-            ->setConfigurationId('main-2')
-            ->setConfiguration(array('x' => 'y'))
-            ->setName('Main2'));
-        $provisioningConfig = (new \Keboola\StorageApi\Options\Components\Configuration())
-            ->setComponentId('provisioning')
-            ->setConfigurationId('main-1')
-            ->setName('Main1');
-        $adminComponentClient->addConfiguration($provisioningConfig);
-
-        // make sure our admin token sees all
-        $allComponents = $adminComponentClient->listComponents();
-        $this->assertCount(2, $allComponents);
-
-        $accessibleComponents = $accessComponentClient->listComponents();
-        $this->assertCount(1, $accessibleComponents);
-        $this->assertEquals($componentAccess[0], $accessibleComponents[0]['id']);
-
-        try {
-            $failComponents = $failComponentClient->listComponents();
-            $this->fail("This token should not be allowed to access components API");
-        } catch (\Keboola\StorageApi\ClientException  $e) {
-        }
-
-        // we should be able to read a config from our accessible component
-        $config = $accessComponentClient->getConfiguration("wr-db", "main-1");
-        $this->assertEquals($config["name"], "Main1");
-
-        // we should be able to add a configuration for our accessible component
-        $config = (new \Keboola\StorageApi\Options\Components\Configuration())
-            ->setComponentId('wr-db')
-            ->setConfigurationId('main-3')
-            ->setName('Main3');
-        $accessComponentClient->addConfiguration($config);
-
-        // check that we can update our configuration
-        $newName = "MAIN-3";
-        $newConfigData = array("foo" => "bar");
-        $config->setName($newName);
-        $config->setConfiguration($newConfigData);
-        $accessComponentClient->updateConfiguration($config);
-        $updatedConfig = $accessComponentClient->getConfiguration($config->getComponentId(), $config->getConfigurationId());
-        $this->assertEquals($updatedConfig['name'], $newName);
-        $this->assertEquals($updatedConfig['configuration'], $newConfigData);
-
-        // we should be able to delete this configuration too
-        $accessComponentClient->deleteConfiguration($config->getComponentId(), $config->getConfigurationId());
-        try {
-            // it should be gone now, and throw a 404
-            $deletedConfig = $accessComponentClient->getConfiguration($config->getComponentId(), $config->getConfigurationId());
-            $this->fail("Configuration should no longer exist, throw a 404");
-        } catch (\Keboola\StorageApi\ClientException  $e) {
-            $this->assertEquals(404, $e->getCode());
-        }
-
-        // we should not be able to add a configuration for any other components
-        try {
-            $accessComponentClient->addConfiguration((new \Keboola\StorageApi\Options\Components\Configuration())
-                ->setComponentId('provisioning')
-                ->setConfigurationId('main-2')
-                ->setConfiguration(array('foo' => 'bar'))
-                ->setName('Main2'));
-            $this->fail("Token was not granted access to this component, should throw an exception");
-        } catch (\Keboola\StorageApi\ClientException  $e) {
-        }
-
-        // tokens should not be able to add configuration rows to configs of components that are inaccessible
-        $configurationRow = new \Keboola\StorageApi\Options\Components\ConfigurationRow($provisioningConfig);
-        $configurationRow->setRowId('main-1-1');
-        try {
-            $accessComponentClient->addConfigurationRow($configurationRow);
-            $this->fail("Token was not granted access to this component, should throw an exception");
-        } catch (\Keboola\StorageApi\ClientException  $e) {
-        }
-
-        // the token with no component access should not be able to add configurations
-        try {
-            $failComponentClient->addConfiguration((new \Keboola\StorageApi\Options\Components\Configuration())
-                ->setComponentId('provisioning')
-                ->setConfigurationId('main-2')
-                ->setConfiguration(array('foo' => 'bar'))
-                ->setName('Main2'));
-            $this->fail("Token was not granted access to this component, should throw an exception");
-        } catch (\Keboola\StorageApi\ClientException  $e) {
-        }
-
-        // let's grant the fail token access to the provisioning component
-        $newFailTokenId = $this->_client->legacyUpdateToken($componentFailTokenId, $bucketPermissions, $description, null, ["provisioning"]);
-        $this->assertEquals($newFailTokenId, $componentFailTokenId);
-        // the fail client should be able to see this component now
-        $nonFailComponents = $failComponentClient->listComponents();
-        $this->assertCount(1, $nonFailComponents);
-        $this->assertEquals("provisioning", $nonFailComponents[0]["id"]);
-
-        // cleanup
-        $this->_clearComponents();
-    }
-
-    public function testAssignNonExistingBucketShouldFail()
-    {
-        $bucketPermissions = array(
-            'out.tohle-je-hodne-blby-nazev' => 'read'
-        );
-
-        try {
-            $this->_client->legacyCreateToken($bucketPermissions, 'Some description');
-            $this->fail('Invalid permissions exception should be thrown');
-        } catch (\Keboola\StorageApi\ClientException $e) {
-            $this->assertEquals('storage.tokens.invalidPermissions', $e->getStringCode());
-        }
-    }
-
-    public function testAllBucketsTokenPermissions()
-    {
-        $description = 'Out read token';
-        $bucketsInitialCount = count($this->_client->listBuckets());
-        $tokenId = $this->_client->legacyCreateToken('manage', $description);
-        $token = $this->_client->getToken($tokenId);
-
-        $client = new \Keboola\StorageApi\Client(array(
-            'token' => $token['token'],
-            'url' => STORAGE_API_URL,
-        ));
-
-        // token getter
-        $this->assertEquals($client->getTokenString(), $token['token']);
-        $this->assertEmpty($token['expires']);
-        $this->assertFalse($token['isExpired']);
-
-        // check assigned buckets
-        $buckets = $client->listBuckets();
-        $this->assertCount($bucketsInitialCount, $buckets);
-
-        // create new bucket with master token
-        $newBucketId = $this->_client->createBucket('test', 'in', 'testing');
-
-        // check if new token has access to token
-        $buckets = $client->listBuckets();
-        $this->assertCount($bucketsInitialCount + 1, $buckets);
-
-        $bucket = $client->getBucket($newBucketId);
-        $client->dropBucket($newBucketId);
-    }
-
-    public function testTokenWithExpiration()
-    {
-        $description = 'Out read token with expiration';
-        $bucketPermissions = array(
-            $this->_outBucketId => 'read'
-        );
-        $twoMinutesExpiration = 2 * 60;
-        $tokenId = $this->_client->legacyCreateToken($bucketPermissions, $description, $twoMinutesExpiration);
-        $token = $this->_client->getToken($tokenId);
-
-        $client = new \Keboola\StorageApi\Client(array(
-            'token' => $token['token'],
-            'url' => STORAGE_API_URL,
-        ));
-        $token = $client->verifyToken();
-        $this->assertNotEmpty($token['expires']);
-        $this->assertFalse($token['isExpired']);
-    }
-
-    public function testExpiredToken()
-    {
-        $description = 'Out read token with expiration';
-        $bucketPermissions =[
-            $this->_outBucketId => 'read',
-        ];
-        $oneSecondExpiration = 1;
-        $tokenId = $this->_client->legacyCreateToken($bucketPermissions, $description, $oneSecondExpiration);
-        $token = $this->_client->getToken($tokenId);
-        $tries = 0;
-
-        $this->expectException(ClientException::class);
-        while ($tries < 7) {
-            $client = new Client([
-                'token' => $token['token'],
-                'url' => STORAGE_API_URL,
-            ]);
-            $client->verifyToken();
-            sleep(pow(2, $tries++));
-        }
-        $this->fail('token should be invalid');
-    }
-
-    public function testTokenUpdateKeepsCanManageBucketsFlag()
-    {
-        $bucketIds = array_map(
-            function ($bucket) {
-                return $bucket['id'];
-            },
-            $this->_client->listBuckets()
-        );
-
-        $this->assertGreaterThan(0, count($bucketIds));
-
-        $tokenId = $this->_client->createToken(
-            'manage',
-            'CanManageBuckets'
-        );
-
-        $token = $this->_client->getToken($tokenId);
-
-        $this->assertTrue($token['canManageBuckets']);
-
-        $this->assertCount(count($bucketIds), $token['bucketPermissions']);
-        $this->assertEmpty(array_diff($bucketIds, array_keys($token['bucketPermissions'])));
-
-        foreach ($token['bucketPermissions'] as $bucketPermission) {
-            $this->assertSame('manage', $bucketPermission);
-        }
-
-        // update token with permissions
-        $this->_client->updateToken(
-            $tokenId,
-            [
-                $this->_outBucketId => 'read',
-            ],
-            'CanManageBuckets update 1'
-        );
-
-        $token = $this->_client->getToken($tokenId);
-
-        $this->assertTrue($token['canManageBuckets']);
-
-        $this->assertCount(count($bucketIds), $token['bucketPermissions']);
-        $this->assertEmpty(array_diff($bucketIds, array_keys($token['bucketPermissions'])));
-
-        foreach ($token['bucketPermissions'] as $bucketPermission) {
-            $this->assertSame('manage', $bucketPermission);
-        }
-
-        // update token without permissions
-        $this->_client->updateToken(
-            $tokenId,
-            [],
-            'CanManageBuckets update 2'
-        );
-
-        $token = $this->_client->getToken($tokenId);
-
-        $this->assertTrue($token['canManageBuckets']);
-
-        $this->assertCount(count($bucketIds), $token['bucketPermissions']);
-        $this->assertEmpty(array_diff($bucketIds, array_keys($token['bucketPermissions'])));
-
-        foreach ($token['bucketPermissions'] as $bucketPermission) {
-            $this->assertSame('manage', $bucketPermission);
-        }
-    }
-
-    public function testTokenWithoutTokensManagePermissionCanListAndViewOnlySelf()
-    {
-        $initialTokens = $this->_client->listTokens();
-
-        $tokenId = $this->_client->createToken([], 'Token without canManageTokens permission');
-
-        $tokens = $this->_client->listTokens();
-        $this->assertCount(count($initialTokens) + 1, $tokens);
-
-        $token = $this->_client->getToken($tokenId);
-        $this->assertFalse($token['canManageTokens']);
-
-        $client = new \Keboola\StorageApi\Client([
-            'token' => $token['token'],
-            'url' => STORAGE_API_URL,
-        ]);
-
-        $verifiedToken = $client->verifyToken();
-
-        $tokens = $client->listTokens();
-        $this->assertCount(1, $tokens);
-
-        $token = reset($tokens);
-        $this->assertSame($verifiedToken['id'], $token['id']);
-
-        $token = $client->getToken($tokenId);
-        $this->assertSame($verifiedToken['id'], $token['id']);
-
-        $tokenId = $this->_client->createToken([], 'Token without canManageTokens permission');
-
-        try {
-            $client->getToken($tokenId);
-            $this->fail('Other token detail with no permissions');
-        } catch (ClientException $e) {
-            $this->assertEquals(403, $e->getCode());
-        }
-    }
-}
